requirements.txt
pylint
pyyaml
yapf

# grpc
grpcio
grpcio-tools
pydantic
betterproto[compiler]

# workflow management
prefect
